Create code 2
Stuff that is in:
First 4 questions, first 4 answers + feedback. VP switcher.

Stuff that doesn't work: 
2 questions have too long answers, Shockonomics is fixing this. VP switcher still only lets you choose 1 VP at max, I'm working on fixing this lol it'll take time. Low priority I think????

Stuff that needs to be added:
Everything else, the rest is just base 2000 (including the answer effects and the issues. Also Perot is getting Nader level polling at game start which is very funny but still something we should hash out a plan for)